RANGER-5655: Consolidate partition-plan helpers to reduce PR file count
Fold PrimaryCatalogRule and bootstrap tests into peers, drop unused PluginScale
DTO, and add a layered review guide to PR #1032.
