--- audit-server/audit-ingestor/src/main/java/org/apache/ranger/audit/producer/kafka/partition/PrimaryCatalogRule.java
@@ -1,31 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.apache.ranger.audit.producer.kafka.partition;
-
-/** One primary {@code auth_to_local} catalog rule and its mapped Kerberos short name. */
-public class PrimaryCatalogRule {
-    final String ruleLine;
-    final String mappedShortName;
-
-    PrimaryCatalogRule(String ruleLine, String mappedShortName) {
-        this.ruleLine        = ruleLine;
-        this.mappedShortName = mappedShortName;
-    }
-}
