+++ audit-server/audit-ingestor/src/test/java/org/apache/ranger/audit/producer/kafka/partition/PartitionPlanBootstrapTest.java
@@ -26,6 +26,7 @@ import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.concurrent.atomic.AtomicReference;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertIterableEquals;
@@ -80,5 +81,114 @@ public class PartitionPlanBootstrapTest {
         assertEquals(15, plan.getTopicPartitionCount());
         assertIterableEquals(List.of(0, 1, 2), plan.getPlugins().get("hdfs").getPartitions());
         assertIterableEquals(List.of(6, 7, 8, 9, 10, 11, 12, 13, 14), plan.getBuffer().getPartitions());
+    }
+
+    @Test
+    public void testBootstrapReturnsExistingPlan() {
+        PartitionPlan existing = samplePlan();
+        InMemoryPartitionPlanRegistry registry = new InMemoryPartitionPlanRegistry(existing);
+        Map<String, Object> config = producerConfig();
+
+        PartitionPlan plan = PartitionPlanBootstrap.bootstrapIfEmpty(registry, TOPIC, config);
+
+        assertEquals(existing, plan);
+        assertEquals(0, registry.getWriteCount());
+    }
+
+    @Test
+    public void testBootstrapPublishesV1WhenEmpty() {
+        InMemoryPartitionPlanRegistry registry = new InMemoryPartitionPlanRegistry(null);
+        Map<String, Object> config = producerConfig();
+
+        PartitionPlan plan = PartitionPlanBootstrap.bootstrapIfEmpty(registry, TOPIC, config);
+
+        assertEquals(1, plan.getVersion());
+        assertEquals(15, plan.getTopicPartitionCount());
+        assertIterableEquals(plan.getPlugins().get("hdfs").getPartitions(), registry.readPlan(TOPIC).getPlugins().get("hdfs").getPartitions());
+        assertEquals(1, registry.getWriteCount());
+    }
+
+    @Test
+    public void testBootstrapAdoptsPeerPlanBeforePublish() {
+        PartitionPlan peerPlan = samplePlan().toBuilder().updatedBy("peer").build();
+        PeerPublishesOnSecondReadRegistry registry = new PeerPublishesOnSecondReadRegistry(peerPlan);
+
+        PartitionPlan plan = PartitionPlanBootstrap.bootstrapIfEmpty(registry, TOPIC, producerConfig());
+
+        assertEquals(peerPlan, plan);
+        assertEquals(0, registry.getWriteCount());
+    }
+
+    private static Map<String, Object> producerConfig() {
+        String propPrefix = AuditServerConstants.PROP_PREFIX_AUDIT_SERVER;
+        Map<String, Object> config = new HashMap<>();
+        config.put(propPrefix + AuditServerConstants.PROP_CONFIGURED_PLUGINS, "hdfs,hiveServer2");
+        config.put(propPrefix + AuditServerConstants.PROP_TOPIC_PARTITIONS_PER_CONFIGURED_PLUGIN, 3);
+        config.put(propPrefix + AuditServerConstants.PROP_BUFFER_PARTITIONS, 9);
+        return config;
+    }
+
+    private static PartitionPlan samplePlan() {
+        return PartitionPlanBootstrap.createInitialPlan(PartitionPlanBootstrapConfig.create(TOPIC, new String[] {"hdfs", "hiveServer2"}, 3, 9));
+    }
+
+    private static final class PeerPublishesOnSecondReadRegistry implements PartitionPlanRegistry {
+        private final PartitionPlan peerPlan;
+        private int readCount;
+        private int writeCount;
+
+        private PeerPublishesOnSecondReadRegistry(PartitionPlan peerPlan) {
+            this.peerPlan = peerPlan;
+        }
+
+        @Override
+        public PartitionPlan readPlan(String auditTopicKey) {
+            readCount++;
+            if (readCount >= 2) {
+                return peerPlan;
+            }
+            return null;
+        }
+
+        @Override
+        public void writePlan(String auditTopicKey, PartitionPlan plan) {
+            writeCount++;
+        }
+
+        @Override
+        public void close() {
+        }
+
+        private int getWriteCount() {
+            return writeCount;
+        }
+    }
+
+    private static class InMemoryPartitionPlanRegistry implements PartitionPlanRegistry {
+        private final AtomicReference<PartitionPlan> planRef;
+        private int writeCount;
+
+        private InMemoryPartitionPlanRegistry(PartitionPlan initialPlan) {
+            this.planRef = new AtomicReference<>(initialPlan);
+        }
+
+        @Override
+        public PartitionPlan readPlan(String auditTopicKey) {
+            return planRef.get();
+        }
+
+        @Override
+        public void writePlan(String auditTopicKey, PartitionPlan plan) {
+            planRef.set(plan);
+            writeCount++;
+        }
+
+        @Override
+        public void close() {
+        }
+
+        private int getWriteCount() {
+            return writeCount;
+        }
     }
 }
