+++ audit-server/audit-ingestor/src/test/java/org/apache/ranger/audit/producer/kafka/partition/PartitionPlanReplacementTest.java
@@ -22,7 +22,6 @@ import org.apache.ranger.audit.producer.kafka.partition.model.OnboardService;
 import org.apache.ranger.audit.producer.kafka.partition.model.PartitionPlan;
 import org.apache.ranger.audit.producer.kafka.partition.model.PartitionPlanReplacement;
 import org.apache.ranger.audit.producer.kafka.partition.model.PluginPartitionAssignment;
-import org.apache.ranger.audit.producer.kafka.partition.model.PluginScale;
 import org.apache.ranger.audit.producer.kafka.partition.model.ServiceAllowlistEntry;
 import org.junit.jupiter.api.Test;
 
@@ -154,6 +153,6 @@ public class PartitionPlanReplacementTest {
     @Test
     public void testRequestValidatorRejectsScaleWithZeroAdditionalPartitions() {
         assertThrows(PartitionPlanException.class,
-                () -> PartitionPlanRequestValidator.validateScalePlugin("hiveServer2", new PluginScale(0, 1)));
+                () -> PartitionPlanRequestValidator.validatePluginScale("hiveServer2", 0, 1));
     }
 }
